Refactor tzdir_set() into own function
I added a warning to make it more when somthing has gone wrong, and removed the special case behaviour for the cctz default directory to make TZDIR more precise.

If we keep seeing tzdir problems, I think we'll need to extract out an exported helper that helps us diagnose the problem.